+++ mithril-stm/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.10.40 (06-29-2026)
+## 0.10.42 (06-29-2026)
+
+### Changed
+
+- Better context for the `IvcCircuitError::CertificateProofRejected` error
+- Better handling of the `GenesisMessagePreimage` in tests
+- Name change for the `AncillaryProverData` variant
+
+## 0.10.41 (06-29-2026)
 
 ### Added